Move types to specific folders
This makes the code base more readable even though we only have a
few classes.
