+++ src/ServiceProviders/StatisticsEventsServiceProvider.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace DDB\Stats;
+namespace DDB\Stats\ServiceProviders;
 
+use DDB\Stats\Events\StatisticsClaimed;
+use DDB\Stats\Listeners\RemoveObsoleteStatistics;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider;
 
 class StatisticsEventsServiceProvider extends EventServiceProvider
